refactor IndirectClient / DirectClient
