+++ pac4j-cas/src/main/java/org/pac4j/cas/client/rest/CasRestBasicAuthClient.java
@@ -46,7 +46,7 @@ public class CasRestBasicAuthClient extends AbstractCasRestClient {
     }
 
     @Override
-    protected void internalInit(final WebContext context) {
+    protected void clientInit(final WebContext context) {
         CommonHelper.assertNotBlank("headerName", this.headerName);
         CommonHelper.assertNotNull("prefixHeader", this.prefixHeader);
