+++ pac4j-cas/src/main/java/org/pac4j/cas/client/rest/CasRestFormClient.java
@@ -44,7 +44,7 @@ public class CasRestFormClient extends AbstractCasRestClient {
     }
 
     @Override
-    protected void internalInit(final WebContext context) {
+    protected void clientInit(final WebContext context) {
         CommonHelper.assertNotBlank("usernameParameter", this.usernameParameter);
         CommonHelper.assertNotBlank("passwordParameter", this.passwordParameter);
