+++ pac4j-core/src/main/java/org/pac4j/core/client/BaseClient.java
@@ -7,8 +7,13 @@ import java.util.List;
 import org.pac4j.core.authorization.generator.AuthorizationGenerator;
 import org.pac4j.core.context.WebContext;
 import org.pac4j.core.credentials.Credentials;
+import org.pac4j.core.credentials.authenticator.Authenticator;
+import org.pac4j.core.credentials.extractor.CredentialsExtractor;
+import org.pac4j.core.exception.CredentialsException;
 import org.pac4j.core.exception.HttpAction;
 import org.pac4j.core.profile.CommonProfile;
+import org.pac4j.core.profile.creator.AuthenticatorProfileCreator;
+import org.pac4j.core.profile.creator.ProfileCreator;
 import org.pac4j.core.util.CommonHelper;
 import org.pac4j.core.util.InitializableWebObject;
 import org.slf4j.Logger;
@@ -22,7 +27,9 @@ import org.slf4j.LoggerFactory;
  * required to initialize the client</li>
  * <li>The name of the client is handled through the {@link #setName(String)} and {@link #getName()} methods</li>
  * <li>After retrieving the user profile, the client can generate the authorization information (roles, permissions and remember-me) by using
- * the appropriate {@link AuthorizationGenerator}.</li>
+ * the appropriate {@link AuthorizationGenerator}</li>
+ * <li>The credentials extraction and validation in the {@link #getCredentials(WebContext)} method are handled by the {@link #credentialsExtractor} and {@link #authenticator} components</li>
+ * <li>The user profile retrieval in the {@link #getUserProfile(Credentials, WebContext)} method is ensured by the {@link #profileCreator} component.</li>
  * </ul>
  *
  * @author Jerome Leleu
@@ -35,6 +42,35 @@ public abstract class BaseClient<C extends Credentials, U extends CommonProfile>
     private String name;
 
     private List<AuthorizationGenerator<U>> authorizationGenerators = new ArrayList<>();
+
+    private CredentialsExtractor<C> credentialsExtractor;
+
+    private Authenticator<C> authenticator;
+
+    private ProfileCreator<C, U> profileCreator = AuthenticatorProfileCreator.INSTANCE;
+
+    /**
+     * Retrieve the credentials.
+     *
+     * @param context the web context
+     * @return the credentials
+     * @throws HttpAction whether an additional HTTP action is required
+     */
+    protected C retrieveCredentials(final WebContext context) throws HttpAction {
+        try {
+            final C credentials = this.credentialsExtractor.extract(context);
+            if (credentials == null) {
+                return null;
+            }
+            this.authenticator.validate(credentials, context);
+            return credentials;
+        } catch (CredentialsException e) {
+            logger.info("Failed to retrieve or validate credentials: {}", e.getMessage());
+            logger.debug("Failed to retrieve or validate credentials", e);
+
+            return null;
+        }
+    }
 
     @Override
     public final U getUserProfile(final C credentials, final WebContext context) throws HttpAction {
@@ -64,7 +100,11 @@ public abstract class BaseClient<C extends Credentials, U extends CommonProfile>
      * @return the user profile
      * @throws HttpAction whether an additional HTTP action is required
      */
-    protected abstract U retrieveUserProfile(final C credentials, final WebContext context) throws HttpAction;
+    protected final U retrieveUserProfile(final C credentials, final WebContext context) throws HttpAction {
+        final U profile = this.profileCreator.create(credentials, context);
+        logger.debug("profile: {}", profile);
+        return profile;
+    }
 
     public void setName(final String name) {
         this.name = name;
@@ -119,9 +159,39 @@ public abstract class BaseClient<C extends Credentials, U extends CommonProfile>
         this.authorizationGenerators.addAll(authorizationGenerators);
     }
 
+    public CredentialsExtractor<C> getCredentialsExtractor() {
+        return credentialsExtractor;
+    }
+
+    public void setCredentialsExtractor(final CredentialsExtractor<C> credentialsExtractor) {
+        if (this.credentialsExtractor == null) {
+            this.credentialsExtractor = credentialsExtractor;
+        }
+    }
+
+    public Authenticator<C> getAuthenticator() {
+        return authenticator;
+    }
+
+    public void setAuthenticator(final Authenticator<C> authenticator) {
+        if (this.authenticator == null) {
+            this.authenticator = authenticator;
+        }
+    }
+
+    public ProfileCreator<C, U> getProfileCreator() {
+        return profileCreator;
+    }
+
+    public void setProfileCreator(final ProfileCreator<C, U> profileCreator) {
+        if (this.profileCreator == null || this.profileCreator == AuthenticatorProfileCreator.INSTANCE) {
+            this.profileCreator = profileCreator;
+        }
+    }
+
     @Override
     public String toString() {
-        return CommonHelper.toString(this.getClass(), "name", getName(),
-                "authorizationGenerators", authorizationGenerators);
+        return CommonHelper.toString(this.getClass(), "name", getName(), "credentialsExtractor", this.credentialsExtractor,
+                "authenticator", this.authenticator, "profileCreator", this.profileCreator, "authorizationGenerators", authorizationGenerators);
     }
 }
