+++ pac4j-core/src/main/java/org/pac4j/core/client/DirectClient.java
@@ -1,31 +1,37 @@
 package org.pac4j.core.client;
 
 import org.pac4j.core.context.WebContext;
-import org.pac4j.core.credentials.extractor.CredentialsExtractor;
-import org.pac4j.core.exception.CredentialsException;
 import org.pac4j.core.exception.HttpAction;
 import org.pac4j.core.exception.TechnicalException;
 import org.pac4j.core.profile.CommonProfile;
-import org.pac4j.core.credentials.authenticator.Authenticator;
 import org.pac4j.core.credentials.Credentials;
-import org.pac4j.core.profile.creator.AuthenticatorProfileCreator;
-import org.pac4j.core.profile.creator.ProfileCreator;
 import org.pac4j.core.redirect.RedirectAction;
 import org.pac4j.core.util.CommonHelper;
 
 /**
- * Direct client type using the {@link CredentialsExtractor}, {@link Authenticator} and {@link ProfileCreator} concepts.
+ * Direct client: credentials are passed and authentication occurs for every HTTP request.
  *
  * @author Jerome Leleu
  * @since 1.9.0
  */
 public abstract class DirectClient<C extends Credentials, U extends CommonProfile> extends BaseClient<C, U> {
 
-    private CredentialsExtractor<C> credentialsExtractor;
+    @Override
+    protected final void internalInit(final WebContext context) {
+        clientInit(context);
 
-    private Authenticator<C> authenticator;
+        // ensures components have been properly initialized
+        CommonHelper.assertNotNull("credentialsExtractor", getCredentialsExtractor());
+        CommonHelper.assertNotNull("authenticator", getAuthenticator());
+        CommonHelper.assertNotNull("profileCreator", getProfileCreator());
+    }
 
-    private ProfileCreator<C, U> profileCreator = AuthenticatorProfileCreator.INSTANCE;
+    /**
+     * Initialize the client.
+     *
+     * @param context the web context
+     */
+    protected abstract void clientInit(WebContext context);
 
     @Override
     public final HttpAction redirect(final WebContext context) throws HttpAction {
@@ -33,73 +39,13 @@ public abstract class DirectClient<C extends Credentials, U extends CommonProfil
     }
 
     @Override
-    public C getCredentials(final WebContext context) throws HttpAction {
+    public final C getCredentials(final WebContext context) throws HttpAction {
         init(context);
-        CommonHelper.assertNotNull("credentialsExtractor", this.credentialsExtractor);
-        CommonHelper.assertNotNull("authenticator", this.authenticator);
-
-        try {
-            final C credentials = this.credentialsExtractor.extract(context);
-            if (credentials == null) {
-                return null;
-            }
-            this.authenticator.validate(credentials, context);
-            return credentials;
-        } catch (CredentialsException e) {
-            logger.info("Failed to retrieve or validate credentials: {}", e.getMessage());
-            logger.debug("Failed to retrieve or validate credentials", e);
-
-            return null;
-        }
-    }
-
-    @Override
-    protected U retrieveUserProfile(final C credentials, final WebContext context) throws HttpAction {
-        CommonHelper.assertNotNull("profileCreator", this.profileCreator);
-
-        final U profile = this.profileCreator.create(credentials, context);
-        logger.debug("profile: {}", profile);
-        return profile;
+        return retrieveCredentials(context);
     }
 
     @Override
     public final RedirectAction getLogoutAction(final WebContext context, final U currentProfile, final String targetUrl) {
         return null;
-    }
-
-    public CredentialsExtractor<C> getCredentialsExtractor() {
-        return credentialsExtractor;
-    }
-
-    public void setCredentialsExtractor(final CredentialsExtractor<C> credentialsExtractor) {
-        if (this.credentialsExtractor == null) {
-            this.credentialsExtractor = credentialsExtractor;
-        }
-    }
-
-    public Authenticator<C> getAuthenticator() {
-        return authenticator;
-    }
-
-    public void setAuthenticator(final Authenticator<C> authenticator) {
-        if (this.authenticator == null) {
-            this.authenticator = authenticator;
-        }
-    }
-
-    public ProfileCreator<C, U> getProfileCreator() {
-        return profileCreator;
-    }
-
-    public void setProfileCreator(final ProfileCreator<C, U> profileCreator) {
-        if (this.profileCreator == null || this.profileCreator == AuthenticatorProfileCreator.INSTANCE) {
-            this.profileCreator = profileCreator;
-        }
-    }
-
-    @Override
-    public String toString() {
-        return CommonHelper.toString(this.getClass(), "name", getName(), "credentialsExtractor", this.credentialsExtractor,
-                "authenticator", this.authenticator, "profileCreator", this.profileCreator);
     }
 }
