+++ pac4j-http/src/main/java/org/pac4j/http/client/direct/DirectBasicAuthClient.java
@@ -29,7 +29,7 @@ public class DirectBasicAuthClient extends DirectClient<UsernamePasswordCredenti
     }
 
     @Override
-    protected void internalInit(final WebContext context) {
+    protected void clientInit(final WebContext context) {
         setCredentialsExtractor(new BasicAuthExtractor(getName()));
     }
 }
