+++ pac4j-http/src/main/java/org/pac4j/http/client/indirect/FormClient.java
@@ -59,9 +59,7 @@ public class FormClient extends IndirectClient<UsernamePasswordCredentials, Comm
     }
 
     @Override
-    protected void internalInit(final WebContext context) {
-        super.internalInit(context);
-
+    protected void clientInit(final WebContext context) {
         CommonHelper.assertNotBlank("loginUrl", this.loginUrl);
         this.loginUrl = callbackUrlResolver.compute(this.loginUrl, context);
         CommonHelper.assertNotBlank("usernameParameter", this.usernameParameter);
