+++ pac4j-oauth/src/main/java/org/pac4j/oauth/client/CasOAuthWrapperClient.java
@@ -40,7 +40,7 @@ public class CasOAuthWrapperClient extends OAuth20Client<CasOAuthWrapperProfile>
     }
     
     @Override
-    protected void internalInit(final WebContext context) {
+    protected void clientInit(final WebContext context) {
         CommonHelper.assertNotBlank("casOAuthUrl", this.casOAuthUrl);
         configuration.setApi(new CasOAuthWrapperApi20(this.casOAuthUrl, this.springSecurityCompliant, this.implicitFlow));
         configuration.setProfileDefinition(new CasOAuthWrapperProfileDefinition());
@@ -48,7 +48,7 @@ public class CasOAuthWrapperClient extends OAuth20Client<CasOAuthWrapperProfile>
         setConfiguration(configuration);
         setLogoutActionBuilder(new CasLogoutActionBuilder<>(casLogoutUrl, "service"));
 
-        super.internalInit(context);
+        super.clientInit(context);
     }
 
     public String getCasOAuthUrl() {
