+++ pac4j-oidc/src/main/java/org/pac4j/oidc/client/KeycloakOidcClient.java
@@ -29,12 +29,12 @@ public class KeycloakOidcClient extends OidcClient<KeycloakOidcProfile> {
     }
 
     @Override
-    protected void internalInit(final WebContext context) {
+    protected void clientInit(final WebContext context) {
         CommonHelper.assertNotNull("configuration", getConfiguration());
         final OidcProfileCreator<KeycloakOidcProfile> profileCreator = new OidcProfileCreator<>(getConfiguration());
         profileCreator.setProfileDefinition(new OidcProfileDefinition<>(x -> new KeycloakOidcProfile()));
         setProfileCreator(profileCreator);
 
-        super.internalInit(context);
+        super.clientInit(context);
     }
 }
